feat: add Checkbox controlled component
