+++ src/components/Common/index.tsx
@@ -1,4 +1,3 @@
-export * from './Checkbox/Checkbox'
 export * from './hooks/useAsyncError'
 export * from './TaxInputs'
 export * from './Button/Button'
@@ -15,3 +14,4 @@ export * from './Inputs/TextField'
 export * from './Inputs/Combobox'
 export * from './Inputs/RadioGroup'
 export * from './Inputs/Select'
+export * from './Inputs/Checkbox'
